test(cli): migrate suite to culture agents noun
