+++ tests/test_agent_argparse_errors.py
@@ -1,13 +1,13 @@
-"""Tests for argparse-style error formatting on culture agent verbs (#333).
+"""Tests for argparse-style error formatting on culture agents verbs (#333).
 
 Two micro-fixes:
 
-* **Item 12** — `culture agent sleep` (and `wake`, etc.) routed
+* **Item 12** — `culture agents sleep` (and `wake`, etc.) routed
   missing-arg errors through `print(..., file=stderr) + sys.exit(1)`
   while every other CLI verb used argparse's standard
   ``<prog>: error: ...`` formatter on stderr with rc 2. Now they all
   match.
-* **Item 11** — `culture agent message` to an unknown nick framed the
+* **Item 11** — `culture agents message` to an unknown nick framed the
   failure as if local config were the source of truth ("Agent 'X' not
   found in config"), which is wrong on a federated mesh. The new
   message points at `culture channel who #general` for a live view.
@@ -22,7 +22,7 @@ import sys
 
 import pytest
 
-from culture.cli import agent
+from culture.cli import agents as agent
 
 
 def _ns(**kw) -> argparse.Namespace:
@@ -53,7 +53,7 @@ def test_sleep_no_args_uses_argparse_error_format(capsys):
 
     assert ei.value.code == 2, "argparse uses rc 2 for usage errors"
     err = capsys.readouterr().err
-    assert "culture agent sleep" in err
+    assert "culture agents sleep" in err
     assert "error:" in err.lower()
     assert "required" in err.lower()
 
@@ -67,7 +67,7 @@ def test_sleep_both_nick_and_all_uses_argparse_error_format(capsys):
 
     assert ei.value.code == 2
     err = capsys.readouterr().err
-    assert "culture agent sleep" in err
+    assert "culture agents sleep" in err
     assert "cannot specify both" in err
 
 
@@ -93,7 +93,7 @@ def test_sleep_unknown_nick_uses_argparse_error_format(capsys):
 
     assert ei.value.code == 2
     err = capsys.readouterr().err
-    assert "culture agent sleep" in err
+    assert "culture agents sleep" in err
     assert "spark-ghost" in err
     assert "not found" in err
 
@@ -104,7 +104,7 @@ def test_sleep_unknown_nick_uses_argparse_error_format(capsys):
 
 
 def test_message_to_unknown_agent_attempts_send_not_local_config_error(monkeypatch, capsys):
-    """`culture agent message` must not short-circuit on local-config absence.
+    """`culture agents message` must not short-circuit on local-config absence.
 
     Issue #333 item 11: the IRC server (with federation) is the source of
     truth for who is reachable on the mesh. The previous behavior errored
@@ -148,9 +148,9 @@ def test_message_to_unknown_agent_attempts_send_not_local_config_error(monkeypat
 
 
 def test_sleep_no_args_exits_2_through_real_cli():
-    """End-to-end: invoking `culture agent sleep` with no args returns rc 2."""
+    """End-to-end: invoking `culture agents sleep` with no args returns rc 2."""
     proc = subprocess.run(
-        [sys.executable, "-m", "culture", "agent", "sleep"],
+        [sys.executable, "-m", "culture", "agents", "sleep"],
         capture_output=True,
         text=True,
         timeout=15,
